+++ CHANGELOG.md
@@ -23,6 +23,8 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * `bList`:
   * `Items.classes` was replaced with `Items.mods`
   * Now `toggleActive` returns a new active item
+  * New public API for `isActive`
+  * Removed deprecated API
 
 * Now the component inherits `iInputText`:
   * `bInput`